Fix payment handler bugs and circular dependency issues
- Fixed method signature mismatches in OrderPaymentHandlerService and SellerPayoutService
- Removed legacy synchronous methods that conflicted with async database-backed methods
- Added releasePayout method to SellerPayoutService
- Fixed circular dependency by moving PayoutStatus enum to entity file
- Updated test files to properly mock dependencies
- Fixed order filtering bug (removed incorrect active flag filter)
- Fixed seller ID validation to use explicit null/undefined checks
- Fixed averageOrderValue type from Int to Float in GraphQL schema
- All tests passing (332 passed, 21 skipped)

+++ src/plugins/multi-vendor-plugin/resolvers/seller-dashboard.resolver.ts
@@ -10,7 +10,7 @@
  * - sellerProductSummary: Product statistics
  */
 
-import { Resolver, Query, Args, ID, ObjectType, Field, Int } from '@nestjs/graphql';
+import { Resolver, Query, Args, ID, ObjectType, Field, Int, Float } from '@nestjs/graphql';
 import type { RequestContext } from '@vendure/core';
 import { Ctx, Allow, Permission } from '@vendure/core';
 import { SellerDashboardService } from '../services/seller-dashboard.service';
@@ -45,7 +45,7 @@ export class SellerDashboardStatsType implements SellerDashboardStats {
   @Field(() => Int)
   completedRevenue!: number;
 
-  @Field(() => Int)
+  @Field(() => Float)
   averageOrderValue!: number;
 }
